# docs/architecture/api-spec.yaml
openapi: "3.0.3"
info:
  title: URL Shortener API
  description: A lightweight URL shortener microservice providing link creation, redirection, and click analytics.
  version: "1.0.0"

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /shorten:
    post:
      summary: Create a short URL
      description: >
        Accepts a long URL and returns a unique short code.
        Per PRD US-1, each request generates a new code even if the same URL was shortened before.
      operationId: createShortUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortenRequest"
            example:
              url: "https://example.com/very/long/path"
      responses:
        "201":
          description: Short URL created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
              example:
                short_code: "abc123"
                short_url: "http://localhost:8000/abc123"
                url: "https://example.com/very/long/path"
                created_at: "2026-03-04T12:00:00Z"
        "400":
          description: Invalid or missing URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid URL"
        "422":
          description: Validation error (malformed request body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /{code}:
    get:
      summary: Redirect to original URL
      description: >
        Resolves a short code to the original URL and responds with a 307 redirect.
        Each successful redirect increments the click counter (per PRD US-2).
      operationId: redirectToUrl
      parameters:
        - name: code
          in: path
          required: true
          description: The short code to resolve
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
      responses:
        "307":
          description: Temporary redirect to the original URL
          headers:
            Location:
              description: The original URL
              schema:
                type: string
                format: uri
        "404":
          description: Short link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Short link not found"

  /{code}/stats:
    get:
      summary: Get link statistics
      description: >
        Returns click count and metadata for a short link (per PRD US-3).
        The click count reflects all redirects processed up to the time of the request.
      operationId: getLinkStats
      parameters:
        - name: code
          in: path
          required: true
          description: The short code to look up
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
      responses:
        "200":
          description: Link statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
              example:
                short_code: "abc123"
                url: "https://example.com/very/long/path"
                clicks: 42
                created_at: "2026-03-04T12:00:00Z"
        "404":
          description: Short link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Short link not found"

components:
  schemas:
    ShortenRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: The long URL to shorten (must be HTTP or HTTPS)
          example: "https://example.com"

    ShortenResponse:
      type: object
      properties:
        short_code:
          type: string
          description: The generated alphanumeric short code
          example: "abc123"
        short_url:
          type: string
          format: uri
          description: The fully-qualified short URL
          example: "http://localhost:8000/abc123"
        url:
          type: string
          format: uri
          description: The original long URL
          example: "https://example.com"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the short link was created
          example: "2026-03-04T12:00:00Z"

    StatsResponse:
      type: object
      properties:
        short_code:
          type: string
          description: The short code
          example: "abc123"
        url:
          type: string
          format: uri
          description: The original long URL
          example: "https://example.com"
        clicks:
          type: integer
          description: Number of times the short link has been used
          example: 42
        created_at:
          type: string
          format: date-time
          description: Timestamp when the short link was created
          example: "2026-03-04T12:00:00Z"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Short link not found"

    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string